requirements.txt
crypto>=1.4.1
Flask>=1.1.2
simplejson>=3.17.2
RPi.GPIO>=0.7.0
google-cloud-pubsub>=1.7.0
requests>=2.24.0
pymongo>=3.11.0
pytz>=2020.1
git+https://github.com/winemug/hbmqtt.git
git+https://github.com/winemug/bluepy.git
